Move all simple document formatting tests to a single file
This commit was purely mechanical, no changes to any test inputs or outputs.
The old file name differentiation made no sense, and its easier to share a single file.
